pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "superduperdb"
description = "🔮 Bring AI to your favourite database 🔮"
readme = "README.md"
license = {file = "LICENSE"}
maintainers = [{name = "SuperDuperDB, Inc.", email = "opensource@superduperdb.com"}]
keywords = [
    "databases",
    "mongodb",
    "data-science",
    "machine-learning",
    "mlops",
    "vector-database",
    "ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: GPU :: NVIDIA CUDA",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
requires-python = ">=3.8"
dynamic = ["version"]
dependencies = [
    "boto3>=1.16",
    "dill>=0.3.6",
    "loguru>=0.7.2",
    "loki-logger-handler>=0.1.1",
    "networkx>=2.8.8",
    "requests>=2.22",  # lower bound from openai and boto3
    "tqdm>=4.64.1",
    "typer>=0.7.0",
    "pylance>=0.6.1,<=0.8.14",
    "pydantic>=1",
    "pymongo>=4.3.3",
    "numpy>=1.24.3",
    "overrides>=7",
    "tenacity>=8.1.0",
    "pillow>=9.4.0",
    "mongomock>=4.1.2",
    "sqlglot<=20.10",
    "PyYAML>=6.0.0",
    "prettytable",
    "python-dotenv",
    "ray[default]>=2.8.1",
    "uvicorn>=0.24.0",
    "fastapi>=0.103.2",
    "ibis-framework>=5.1.0",
    "sqlalchemy>=2.0.28",
    "python-magic"
]

[project.urls]
homepage = "https://www.superduperdb.com/"
documentation = "https://docs.superduperdb.com"
source = "https://github.com/SuperDuperDB/superduperdb/"

[tool.setuptools.packages.find]
include = ["superduperdb*"]

[tool.setuptools.dynamic]
version = {attr = "superduperdb.__version__"}

[tool.black]
skip-string-normalization = true
target-version = ["py38"]

[tool.mypy]
ignore_missing_imports = true
no_implicit_optional = true
warn_unused_ignores = true
disable_error_code = ["has-type", "attr-defined", "assignment", "misc", "override"]

[tool.pytest.ini_options]
addopts = "-W ignore"

[tool.interrogate]
color = true
exclude = []
fail-under = 30.1
ignore-magic = true
ignore-nested-classes = false
ignore-nested-functions = true
ignore-private = true
ignore-property-decorators = true
ignore-regex = []
ignore-semiprivate = true
omit-covered-files = true
quiet = false
verbose = 0
whitelist-regex = []

[tool.ruff]
extend-select = [
    "I", # Missing required import (auto-fixable)
    "F", # PyFlakes
    #"W", # PyCode Warning
    "E", # PyCode Error
    #"N", # pep8-naming
    "D", # pydocstyle
]
ignore = [
  "D100", # Missing docstring in public module
  "D104", # Missing docstring in public package
  "D107", # Missing docstring in __init__
  "D105", # Missing docstring in magic method
  "D212", # Multi-line docstring summary should start at the first line
  "D213", # Multi-line docstring summary should start at the second line
  "D401",
  "E402",
]

[tool.ruff.isort]
combine-as-imports = true

[tool.ruff.per-file-ignores]
"test/**" = ["D"]
